feat: "`date` function with slider"
