+++ shelter/pages/main/index.js
@@ -1,6 +1,6 @@
 import { nextBtn, handle, loadedCards, prevBtn, returnedCountCards } from "../../js/infinite-slider-carousel.js";
-import { createdMenuBurger,clickedBurger,BTN_BURGER } from "../../js/burger.js";
-import { getModalWindow } from "../../js/modal-window.js";
+import { createdMenuBurger, clickedBurger, BTN_BURGER } from "../../js/burger.js";
+import { showModalToClickedCard,closemodal } from "../../js/modal-window.js";
 
 //*____________________________________________________//
 
@@ -13,21 +13,10 @@ export const pets = await fetch("../../assets/data/pets.json")
     return dta;
   })
   .catch((error) => console.log(error));
-console.log(pets);
-
-console.log('dasda');
-
-
-
-
-
-
 
 
 if (document.readyState == "complete") createdMenuBurger();
 BTN_BURGER.addEventListener("click", clickedBurger);
-
-
 
 loadedCards(); // added pet cards to dom
 
@@ -35,41 +24,15 @@ window.addEventListener("resize", () => {
   document.querySelector(".block-card__slider").innerHTML = "";
   loadedCards();
   console.log(returnedCountCards(document.querySelector(".block-card__slider").parentNode));
-
-
 }); // changes the number of cards when changing the size of the window
 nextBtn.addEventListener("click", handle);
 prevBtn.addEventListener("click", handle);
-
-
-
-
-
+//* when clicked to card ---------------------------------------------------------------------------------
 document.body.querySelectorAll(".pets__card").forEach((card) => {
-
-
   card.addEventListener("click", showModalToClickedCard);
-
 });
-function showModalToClickedCard (ev)  {
-  const click = ev.target;
-  let nameCard;
-  if (click.classList.contains("card__btn")) {
-    this.removeListener("click", showModalToClickedCard);
-  } else if (click.closest(".pets__card") && !click.classList.contains("pets__card")) {
-    nameCard = click.parentNode.id;
-  } else if (click.classList.contains("pets__card")) {
-    console.log(click.id);
-    nameCard = click.id;
-  }
-  let clikedCard = pets.find((e) => {
-    if (e.name === nameCard) {
-      return e;
-    }
-  });
-  getModalWindow(clikedCard).showModal();
-  console.log("click card");
 
-  document.body.classList.add("stopScroll");
-};
+
+
+
