+++ shelter/scss/style.css
@@ -1700,7 +1700,8 @@ body.page-pets .btn-number {
 .modal-window {
   max-width: 952px;
   width: 74.375%;
-  height: 500px;
+  height: -moz-max-content;
+  height: max-content;
   background: transparent;
   border: none;
   padding: 0;
@@ -1709,6 +1710,13 @@ body.page-pets .btn-number {
   .modal-window {
     width: 88.8020833333%;
     height: 402px;
+  }
+}
+@media (max-width: 320px) {
+  .modal-window {
+    width: 84.6875%;
+    display: flex;
+    flex-direction: column;
   }
 }
 .modal-window .btn-modal {
@@ -1723,7 +1731,7 @@ body.page-pets .btn-number {
   width: 94.5378151261%;
   display: flex;
   flex-direction: row;
-  justify-content: space-around;
+  justify-content: space-between;
   align-items: center;
   background: #fafafa;
   border-radius: 0.5625rem;
@@ -1732,6 +1740,11 @@ body.page-pets .btn-number {
 @media (max-width: 768px) {
   .modal-window .modal-container {
     width: 92.3753665689%;
+  }
+}
+@media (max-width: 320px) {
+  .modal-window .modal-container {
+    padding: 0.625rem;
   }
 }
 .modal-window .modal-container .modal-img {
@@ -1745,6 +1758,11 @@ body.page-pets .btn-number {
     height: 350px;
   }
 }
+@media (max-width: 320px) {
+  .modal-window .modal-container .modal-img {
+    display: none;
+  }
+}
 .modal-window .modal-container .modal-content {
   display: flex;
   flex-direction: column;
@@ -1755,6 +1773,17 @@ body.page-pets .btn-number {
 @media (max-width: 768px) {
   .modal-window .modal-container .modal-content {
     gap: 1.25rem;
+  }
+}
+@media (max-width: 320px) {
+  .modal-window .modal-container .modal-content {
+    width: 100%;
+  }
+}
+@media (max-width: 320px) {
+  .modal-window .modal-container .modal-content {
+    gap: 1.5rem;
+    width: 100%;
   }
 }
 .modal-window .modal-container .modal-content .modal-description {
@@ -1775,14 +1804,29 @@ body.page-pets .btn-number {
   font-size: 2.1875rem;
   line-height: 130%;
 }
+@media (max-width: 320px) {
+  .modal-window .modal-container .modal-content .modal-description .pet-name {
+    text-align: center;
+  }
+}
 .modal-window .modal-container .modal-content .modal-description .pet-typeBreed {
   font-size: 1.25rem;
   line-height: 115%;
+}
+@media (max-width: 320px) {
+  .modal-window .modal-container .modal-content .modal-description .pet-typeBreed {
+    text-align: center;
+  }
 }
 .modal-window .modal-container .modal-content .modal-description .pet-description {
   margin-top: 1.875rem;
   font-size: 0.9375rem;
   line-height: 110%;
+}
+@media (max-width: 768px) {
+  .modal-window .modal-container .modal-content .modal-description .pet-description {
+    margin-top: 1.25rem;
+  }
 }
 @media (max-width: 768px) {
   .modal-window .modal-container .modal-content .modal-description .pet-description {
